Rearrange tests into their own test files
